requirements.txt
fastapi
uvicorn
python-dotenv
httpx
bcrypt
sqlalchemy
psycopg2-binary
